# docs/swagger.yaml
definitions:
  command.SwaggerCreate:
    properties:
      description:
        example: describe something
        type: string
      https_command:
        $ref: '#/definitions/https_command.SwaggerCreate'
      name:
        example: test
        type: string
      protocol:
        example: http
        type: string
    required:
    - description
    - https_command
    - name
    - protocol
    type: object
  command.SwaggerResponse:
    properties:
      created_at:
        example: "2022-08-15T06:36:36Z"
        type: string
      description:
        example: describe something
        type: string
      https_command:
        $ref: '#/definitions/https_command.SwaggerResponse'
      id:
        example: 1
        type: integer
      name:
        example: test
        type: string
      protocol:
        example: http
        type: string
      updated_at:
        example: "2022-08-15T06:36:36Z"
        type: string
    required:
    - created_at
    - description
    - https_command
    - id
    - name
    - protocol
    - updated_at
    type: object
  command.SwaggerUpdate:
    properties:
      description:
        example: describe something
        type: string
      https_command:
        $ref: '#/definitions/https_command.SwaggerUpdate'
      name:
        example: test
        type: string
      protocol:
        example: http
        type: string
    type: object
  header_template.SwaggerCreate:
    properties:
      data:
        type: object
      name:
        type: string
    required:
    - data
    - name
    type: object
  header_template.SwaggerResponse:
    properties:
      data:
        type: object
      id:
        example: 1
        type: integer
      name:
        type: string
    required:
    - data
    - id
    - name
    type: object
  header_template.SwaggerUpdate:
    properties:
      data:
        type: object
      name:
        type: string
    type: object
  https_command.SwaggerCreate:
    properties:
      authorization_type:
        example: token
        type: string
      body:
        type: object
      body_type:
        example: json
        type: string
      header:
        type: object
      method:
        example: POST
        type: string
      url:
        example: http://localhost:9800/api
        type: string
    required:
    - method
    - url
    type: object
  https_command.SwaggerResponse:
    properties:
      authorization_type:
        example: token
        type: string
      body:
        type: object
      body_type:
        example: json
        type: string
      command_id:
        example: 1
        type: integer
      header:
        type: object
      id:
        example: 1
        type: integer
      method:
        example: "1"
        type: string
      url:
        example: http://localhost:9800/api
        type: string
    required:
    - authorization_type
    - body
    - body_type
    - command_id
    - header
    - id
    - method
    - url
    type: object
  https_command.SwaggerUpdate:
    properties:
      authorization_type:
        example: token
        type: string
      body:
        type: object
      body_type:
        example: json
        type: string
      header:
        type: object
      method:
        example: POST
        type: string
      url:
        example: http://localhost:9800/api
        type: string
    type: object
...
host: localhost:5487
info:
  contact:
    email: wwilson008@gmail.com
    name: Wilson
    url: https://github.com/littlebluewhite
  description: This is a schedule-command server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Schedule-Command swagger API
  version: "1.0"
paths:
  /command/api:
    get:
      description: Get all commands
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/command.SwaggerResponse'
            type: array
      summary: Show all commands
      tags:
      - command
    post:
      consumes:
      - application/json
      parameters:
      - description: command body
        in: body
        name: command
        required: true
        schema:
          $ref: '#/definitions/command.SwaggerCreate'
      produces:
      - application/json
      responses:
        "200":
          description: created success
          schema:
            type: string
      summary: Create commands
      tags:
      - command
  /command/api/{id}:
    delete:
      parameters:
      - description: command id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: delete successfully
          schema:
            type: string
      summary: Delete commands
      tags:
      - command
    get:
      description: Get commands by id
      parameters:
      - description: command id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/command.SwaggerResponse'
      summary: Show commands
      tags:
      - command
    patch:
      consumes:
      - application/json
      parameters:
      - description: command id
        in: path
        name: id
        required: true
        type: integer
      - description: modify command body
        in: body
        name: command
        required: true
        schema:
          $ref: '#/definitions/command.SwaggerUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/command.SwaggerResponse'
      summary: Update commands
      tags:
      - command
  /header_template/api:
    get:
      description: Get all header templates
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/header_template.SwaggerResponse'
            type: array
      summary: Show all header templates
      tags:
      - header_template
    post:
      consumes:
      - application/json
      parameters:
      - description: header template body
        in: body
        name: header_template
        required: true
        schema:
          $ref: '#/definitions/header_template.SwaggerCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create header templates
      tags:
      - header_template
  /header_template/api/{id}:
    delete:
      parameters:
      - description: header template id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete header templates
      tags:
      - header_template
    get:
      description: Get header templates by id
      parameters:
      - description: header template id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/header_template.SwaggerResponse'
      summary: Show header templates
      tags:
      - header_template
    patch:
      consumes:
      - application/json
      parameters:
      - description: header template id
        in: path
        name: id
        required: true
        type: integer
      - description: modify header template body
        in: body
        name: header_template
        required: true
        schema:
          $ref: '#/definitions/header_template.SwaggerUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/header_template.SwaggerResponse'
      summary: Update header templates
      tags:
      - header_template
...
swagger: "2.0"